Add support for morsayati's reckoning
- Add a method to lookup DungeonType in the plugin
- When it is a raid, give each player 4 characters
- Move DungeonTypes enum to Photon.Shared
