+++ DragaliaAPI.Photon.Shared/Enums/DungeonTypes.cs
@@ -1,0 +1,31 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace DragaliaAPI.Photon.Shared.Enums
+{
+    public enum DungeonTypes
+    {
+        None,
+        Normal,
+        Raid,
+        Random,
+        Field,
+        Versus,
+        Wave,
+        Defense,
+        Tutorial = 9,
+        RaidSingle,
+        DragonSolo,
+        DragonRelay,
+        BattleRoyal,
+        NormalSolo,
+        PartySwitch,
+        Walker,
+        Scoring,
+        RaidSixteen,
+        Dmode
+    }
+}
