+++ DragaliaAPI.Photon.Shared/Enums/MatchingTypes.cs
@@ -2,7 +2,7 @@
 using System.Collections.Generic;
 using System.Text;
 
-namespace DragaliaAPI.Photon.Shared
+namespace DragaliaAPI.Photon.Shared.Enums
 {
     public enum MatchingTypes
     {
