+++ DragaliaAPI.Photon.Shared/Requests/GameModifyMatchingTypeRequest.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Text;
+using DragaliaAPI.Photon.Shared.Enums;
 
 namespace DragaliaAPI.Photon.Shared.Requests
 {
