+++ DragaliaAPI.Photon.StateManager.Test/Event/MatchingTypeTest.cs
@@ -5,7 +5,7 @@ using DragaliaAPI.Photon.StateManager.Models;
 using DragaliaAPI.Photon.StateManager.Test.Helpers;
 using System.Net.Http.Json;
 using Xunit.Abstractions;
-using DragaliaAPI.Photon.Shared;
+using DragaliaAPI.Photon.Shared.Enums;
 
 namespace DragaliaAPI.Photon.StateManager.Test.Event;
 
@@ -35,7 +35,7 @@ public class MatchingTypeTest : TestFixture
                 RoomId = 12345,
                 Name = "f162d896-f59e-416c-8df2-46a7649e1074",
                 MatchingCompatibleId = 36,
-                MatchingType = Shared.MatchingTypes.Anyone,
+                MatchingType = MatchingTypes.Anyone,
                 QuestId = 301010103,
                 StartEntryTime = DateTimeOffset.UtcNow,
                 EntryConditions = new()
