+++ DragaliaAPI.Photon.StateManager.Test/Get/GameListTest.cs
@@ -5,6 +5,7 @@ using System.Net;
 using System.Net.Http.Json;
 using System.Text;
 using System.Threading.Tasks;
+using DragaliaAPI.Photon.Shared.Enums;
 using DragaliaAPI.Photon.Shared.Models;
 using DragaliaAPI.Photon.StateManager.Models;
 using Xunit.Abstractions;
@@ -30,7 +31,7 @@ public class GameListTest : TestFixture
                     RoomId = 12345,
                     Name = "639d263a-e05a-4432-952f-fa941e7f7f40",
                     MatchingCompatibleId = 36,
-                    MatchingType = Shared.MatchingTypes.Anyone,
+                    MatchingType = MatchingTypes.Anyone,
                     QuestId = 301010103,
                     StartEntryTime = DateTimeOffset.UtcNow,
                     EntryConditions = new()
@@ -55,7 +56,7 @@ public class GameListTest : TestFixture
                     RoomId = 12345,
                     Name = "0023679c-c449-45bd-881d-3e29a9d982ac",
                     MatchingCompatibleId = 36,
-                    MatchingType = Shared.MatchingTypes.ById,
+                    MatchingType = MatchingTypes.ById,
                     QuestId = 301010103,
                     StartEntryTime = DateTimeOffset.UtcNow,
                     EntryConditions = new()
@@ -98,7 +99,7 @@ public class GameListTest : TestFixture
                     RoomId = 12345,
                     Name = "d4592641-6f07-46d9-8b45-21059b27d1e2",
                     MatchingCompatibleId = 36,
-                    MatchingType = Shared.MatchingTypes.Anyone,
+                    MatchingType = MatchingTypes.Anyone,
                     QuestId = 301010103,
                     StartEntryTime = DateTimeOffset.UtcNow,
                     EntryConditions = new()
@@ -122,7 +123,7 @@ public class GameListTest : TestFixture
                     RoomId = 12345,
                     Name = "c7c494cc-0ffc-46d6-a19b-cbfcad903225",
                     MatchingCompatibleId = 36,
-                    MatchingType = Shared.MatchingTypes.Anyone,
+                    MatchingType = MatchingTypes.Anyone,
                     QuestId = 219010102,
                     StartEntryTime = DateTimeOffset.UtcNow,
                     EntryConditions = new()
