+++ DragaliaAPI.Photon.StateManager/Controllers/EventController.cs
@@ -1,4 +1,4 @@
-using DragaliaAPI.Photon.Shared;
+using DragaliaAPI.Photon.Shared.Enums;
 using DragaliaAPI.Photon.Shared.Models;
 using DragaliaAPI.Photon.Shared.Requests;
 using DragaliaAPI.Photon.StateManager.Authentication;
