+++ DragaliaAPI.Photon.StateManager/Models/RedisGame.cs
@@ -1,4 +1,4 @@
-﻿using DragaliaAPI.Photon.Shared;
+﻿using DragaliaAPI.Photon.Shared.Enums;
 using DragaliaAPI.Photon.Shared.Models;
 using DragaliaAPI.Photon.StateManager.Controllers;
 using Newtonsoft.Json;
