+++ DragaliaAPI.Shared.Test/Unit/MasterAssetTest.cs
@@ -1,4 +1,5 @@
 ﻿using System.Reflection.Emit;
+using DragaliaAPI.Photon.Shared.Enums;
 using DragaliaAPI.Shared.Definitions.Enums;
 using DragaliaAPI.Shared.MasterAsset;
 using DragaliaAPI.Shared.MasterAsset.Models;
