+++ DragaliaAPI.Shared/DragaliaAPI.Shared.csproj
@@ -13,6 +13,10 @@
 	</ItemGroup>
 
 	<ItemGroup>
+	  <ProjectReference Include="..\DragaliaAPI.Photon.Shared\DragaliaAPI.Photon.Shared.csproj" />
+	</ItemGroup>
+
+	<ItemGroup>
 		<None Update="Resources\AbilityCrest.json">
 		  <CopyToOutputDirectory>PreserveNewest</CopyToOutputDirectory>
 		</None>
