+++ DragaliaAPI.Shared/MasterAsset/Models/QuestData.cs
@@ -1,4 +1,5 @@
-﻿using DragaliaAPI.Shared.Definitions.Enums;
+﻿using DragaliaAPI.Photon.Shared.Enums;
+using DragaliaAPI.Shared.Definitions.Enums;
 
 namespace DragaliaAPI.Shared.MasterAsset.Models;
